♻️ refactor(hypothesis): make_nonnegative public; drop astro's private _src reach-in
Closes #550. The `coordinaxs.hypothesis.astro` strategy tree (shipped by the
astro distribution) imported `make_nonnegative` from
`coordinaxs.hypothesis.distances._src.utils` — a private module of a *different*
distribution.

Investigation showed the layout itself is correct, not the coupling: the base
`coordinaxs.hypothesis` distribution is deliberately astro-agnostic (it does not
depend on `coordinaxs.astro`), so astro's own strategies rightly live *with*
astro (the dual tree, wired via astro's `[hypothesis]` extra). Moving the tree
into the hypothesis distribution — the other option floated in #550 — would
invert that clean dependency and force astro onto every hypothesis user. So the
real fix is just to stop reaching into another distribution's `_src`.

`make_nonnegative` already has an `__all__` and a full docstring, so promote it
to the public API of `coordinaxs.hypothesis.distances` and import the public
name from astro's `plx.py`. No behavior change; the astro wheel still ships both
trees with no namespace `__init__` leak, and 45 astro/hypothesis-distances tests
pass.
