+++ packages/coordinaxs.hypothesis/src/coordinaxs/hypothesis/distances/_src/__init__.py
@@ -1,5 +1,6 @@
 """Hypothesis strategies for Distance quantities."""
 
-__all__ = ("distances",)
+__all__ = ("distances", "make_nonnegative")
 
 from .dist import distances
+from .utils import make_nonnegative
